Add accessibility improvements and modular structure
